pyproject.toml
[project]
name = "gestao-ti"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-material-admin (>=1.8.6,<2.0.0)",
    "django-import-export (>=4.3.5,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
